+++ src/common/const.tsx
@@ -1,0 +1,1 @@
+export const DRAWER_WIDTH = 240
